going back and forth between pages, still working on it but testing if this will work
