--- src/LoginOrCreate.form
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="LoginOrCreate">
-  <grid id="27dc6" binding="LoginOrCreatePanel" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="504" height="400"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <component id="abdd1" class="javax.swing.JButton" binding="haveAnAccountLoginButton">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="0" use-parent-layout="false">
-            <preferred-size width="185" height="41"/>
-          </grid>
-        </constraints>
-        <properties>
-          <text value="Have an account? Login here"/>
-        </properties>
-      </component>
-      <component id="38ffb" class="javax.swing.JButton" binding="createAccountButton" default-binding="true">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="0" use-parent-layout="false">
-            <preferred-size width="185" height="41"/>
-          </grid>
-        </constraints>
-        <properties>
-          <text value="Create Account"/>
-        </properties>
-      </component>
-    </children>
-  </grid>
-</form>
