+++ src/com/company/Main.java
@@ -3,9 +3,7 @@ package com.company;
 import javax.swing.*;
 
 public class Main {
-
     public static void main(String[] args) {
-        User user = new User();
-        new LoginFrame(user);
+        new FrontPage();
     }
 }
